requirements.txt
openai
google-cloud-firestore
pinecone-client
firebase-admin
youtube-search-python
rake-nltk